pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mteam-mcp"
version = "0.1.0"
description = "MCP server for M-Team (馒头) private torrent tracker - search, detail, and download"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "07freedom" }
]
keywords = ["mcp", "mteam", "torrent", "fastmcp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet",
]
dependencies = [
    "fastmcp>=2.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "hatch",
]

[project.scripts]
mteam-mcp = "mteam_mcp.server:main"

[project.urls]
Homepage = "https://github.com/07freedom/mteam-mcp"
Repository = "https://github.com/07freedom/mteam-mcp"
Documentation = "https://github.com/07freedom/mteam-mcp#readme"

[tool.hatch.build.targets.wheel]
packages = ["mteam_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "mteam_mcp/**",
    ".env_example",
]
